+++ LibMagestore/src/main/java/com/magestore/app/lib/model/sales/OrderItemParams.java
@@ -13,9 +13,9 @@ public interface OrderItemParams extends Model {
 
     String getOrderItemId();
 
-    int getQty();
+    float getQty();
 
-    void setQty(int strQty);
+    void setQty(float strQty);
 
     String getAdditionalData();
