+++ LibMagestore/src/main/java/com/magestore/app/pos/service/checkout/POSCartService.java
@@ -145,7 +145,7 @@ public class POSCartService extends AbstractService implements CartService {
      * @return
      */
     @Override
-    public CartItem create(Product product, int quantity, float price) {
+    public CartItem create(Product product, float quantity, float price) {
         CartItem cartItem = new PosCartItem();
         cartItem.setTypeNormal();
         cartItem.setProduct(product);
@@ -163,7 +163,7 @@ public class POSCartService extends AbstractService implements CartService {
     }
 
     @Override
-    public CartItem create(Product product, int quantity) {
+    public CartItem create(Product product, float quantity) {
         return create(product, quantity, product.getFinalPrice());
     }
 
@@ -429,8 +429,8 @@ public class POSCartService extends AbstractService implements CartService {
     }
 
     @Override
-    public void increase(CartItem cartItem, int quantity) {
-        int newQuantity = cartItem.getQuantity() + quantity;
+    public void increase(CartItem cartItem, float quantity) {
+        float newQuantity = cartItem.getQuantity() + quantity;
         newQuantity = newQuantity > 0 ? newQuantity : cartItem.getProduct().getQuantityIncrement();
         float newPrice = cartItem.getUnitPrice() * newQuantity;
         cartItem.setQuantity(newQuantity);
@@ -443,8 +443,8 @@ public class POSCartService extends AbstractService implements CartService {
     }
 
     @Override
-    public void substract(CartItem cartItem, int quantity) {
-        int newQuantity = cartItem.getQuantity() - quantity;
+    public void substract(CartItem cartItem, float quantity) {
+        float newQuantity = cartItem.getQuantity() - quantity;
         newQuantity = newQuantity > 0 ? newQuantity : cartItem.getProduct().getQuantityIncrement();
         float newPrice = cartItem.getUnitPrice() * newQuantity;
         cartItem.setQuantity(newQuantity);
@@ -452,7 +452,7 @@ public class POSCartService extends AbstractService implements CartService {
     }
 
     @Override
-    public CartItem insert(Checkout checkout, String productID, String productName, int quantity, float price) throws IOException, InstantiationException, ParseException, IllegalAccessException {
+    public CartItem insert(Checkout checkout, String productID, String productName, float quantity, float price) throws IOException, InstantiationException, ParseException, IllegalAccessException {
         // nếu chưa có đơn hàng, bo qua
         if (checkout == null) return null;
 
@@ -479,7 +479,7 @@ public class POSCartService extends AbstractService implements CartService {
      * @param price
      */
     @Override
-    public CartItem insert(Checkout checkout, Product product, int quantity, float price) throws InstantiationException, IllegalAccessException, ParseException, IOException, ServiceException {
+    public CartItem insert(Checkout checkout, Product product, float quantity, float price) throws InstantiationException, IllegalAccessException, ParseException, IOException, ServiceException {
         // nếu chưa có đơn hàng, bo qua
         if (checkout == null) return null;
 
@@ -514,7 +514,7 @@ public class POSCartService extends AbstractService implements CartService {
         // có rồi thì cập nhật lại số lượng
         else {
             // tính toán số lượng mới
-            int newQuantity = cartItem.getQuantity() + quantity;
+            float newQuantity = cartItem.getQuantity() + quantity;
             // kiểm tra số lượng trước có đủ trong kho không
             if (!validateStock(checkout, product, newQuantity)) throw new ServiceException("Not enough quantity");
             // cập nhật số lượng
@@ -543,7 +543,7 @@ public class POSCartService extends AbstractService implements CartService {
      * @param quantity
      */
     @Override
-    public CartItem insert(Checkout checkout, Product product, int quantity) throws IOException, InstantiationException, ParseException, IllegalAccessException {
+    public CartItem insert(Checkout checkout, Product product, float quantity) throws IOException, InstantiationException, ParseException, IllegalAccessException {
         return insert(checkout, product, quantity, product.getFinalPrice());
     }
 
@@ -613,7 +613,7 @@ public class POSCartService extends AbstractService implements CartService {
         // đã có item, giảm trừ số lượng xuống nhưng không thể ít hơn 1
         if (cartItem != null) {
             // Tính số lượng item mới
-            int newQuantity = cartItem.getQuantity() - subQuantity;
+            float newQuantity = cartItem.getQuantity() - subQuantity;
             if (newQuantity < 0) newQuantity = 1;
 
             // Cập nhật số lượng mới
@@ -714,7 +714,7 @@ public class POSCartService extends AbstractService implements CartService {
      * @return
      */
     @Override
-    public boolean validateStock(Checkout checkout, Product product, int quantity) {
+    public boolean validateStock(Checkout checkout, Product product, float quantity) {
         if (!product.isInStock()) return false;
         if (quantity < product.getAllowMinQty()) return false;
         if (quantity > product.getAllowMaxQty() && product.getAllowMaxQty() > product.getAllowMinQty()) return false;
